Minor refactor of file route function location
