--- functions/src/api/routes/sponsors/sponsors.http
@@ -1,10 +1,0 @@
-POST http://127.0.0.1:5001/conferencecenterr/us-central1/api/v1/sponsors
-#POST localhost:3000/v1/sponsors
-Content-Type: application/json
-
-{
-    "id": "slug-sponsor-name",
-    "name": "Sponsor Name",
-    "logoUrl": "https://example.com",
-    "website": "https://example.com"
-}
